% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live_markets.R
\name{get_live_sports}
\alias{get_live_sports}
\title{Get all in-play sports markets}
\usage{
get_live_sports()
}
\value{
returns a dataframe of all in-play bets available currently
}
\description{
Returns all currently running in-play betting markets
}
\examples{
\dontrun{
try({
df <- get_live_sports()
})
}
}
