% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_race_data.R
\name{get_past_races}
\alias{get_past_races}
\title{Get past race content}
\usage{
get_past_races(meet_date, venue_mnem, race_type, race_num = NULL)
}
\arguments{
\item{meet_date}{race meet date in 'YYYY-MM-DD' format}

\item{venue_mnem}{the mnemonic of the track}

\item{race_type}{the type of race, either R, H, G}

\item{race_num}{integer if the race number. If null,  then all races for meet(s) returned}
}
\value{
returns a list of a race's data including runners, markets, dividends and pools
}
\description{
Returns a list of a race or many races data including runners, markets, dividends and pools
Can either use this function and provide it the race API URL, or use \code{get_past_races}
to get the same results. Note this can return results for many race meets.
Can then use the parsing functions to get the data you want from these lists.
}
\examples{
\dontrun{
try({
out <- get_past_races(meet_date = c('2022-09-03', '2022-09-10'),
                     venue_mnem = 'M', race_type = 'R')
})
}
}
