% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sports_market.R
\name{get_sports_market}
\alias{get_sports_market}
\title{Get all markets fr selected competition}
\usage{
get_sports_market(competition_name)
}
\arguments{
\item{competition_name}{the name of the TAB competition. Found under the competitions page selector}
}
\value{
returns a dataframe of bets available for the selected competition
}
\description{
Returns all betting markets data for a selected TAB competition
}
\examples{
\dontrun{
try({
df <- get_sports_market("Major League Baseball Futures")
})
}
}
