% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_race_meet_meta}
\alias{load_race_meet_meta}
\title{Load race meet metadata}
\usage{
load_race_meet_meta(cal_year)
}
\arguments{
\item{cal_year}{calendar year to get data for}
}
\value{
returns a dataframe all race meets meta data for a selected calendar year
}
\description{
Loading version of \code{get_race_meet_meta}
Returns all race meets meta data for a whole calendar year
}
\examples{
\donttest{
try({
df <- load_race_meet_meta(cal_year=2022)
})
}
}
