% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_race_data.R
\name{parse_pools}
\alias{parse_pools}
\title{Parse race betting pools}
\usage{
parse_pools(race_list)
}
\arguments{
\item{race_list}{output from \code{get_past_race_content} or \code{get_past_races}}
}
\value{
returns a dataframe of all race meet betting pools
}
\description{
Returns a data frame of a race's betting pools
A function used to parse output race lists from \code{get_past_race_content}
or \code{get_past_races}
}
\examples{
\dontrun{
try({
races <- get_race_data(meet_date="2022-09-03", venue_mnem="M", race_type="R", race_num=NULL)
df <- parse_pools(race_list=races)
})
}
}
